Moved to Create resource idiom ResourceId now a seperated enum from uint32_t added size and alignment to bundle handlers Tidy up of Image creation Render/types tidied shaderc library added and library working

+++ libs/vulkan/commandqueue.cpp
@@ -63,7 +63,7 @@ auto CommandQueue::submit(std::shared_ptr<Render::Fence> const& fence_) -> void
 		if(encoder->endSemaphore) cbEndSemaphores.push_back(encoder->endSemaphore->vulkanSemaphore);
 	}
 
-	submitInfo.commandBufferCount = commandBuffers.size();
+	submitInfo.commandBufferCount = (uint32_t) commandBuffers.size();
 	submitInfo.pCommandBuffers = commandBuffers.data();
 
 	submitInfo.waitSemaphoreCount = cbBeginSemaphores.size();
